+++ electron-starter.babel.ts
@@ -1,14 +1,40 @@
-import { BrowserWindow, app } from "electron";
+import { BrowserWindow, app, Menu, ipcMain } from "electron";
 const path = require("path");
-
 const AppUrl = "http://localhost:1500/";
 
 declare global {
   interface Window {
     ipcMain: Electron.IpcMain;
     ipcRenderer: Electron.IpcRenderer;
+    appLanguage: string;
   }
 }
+const menu = Menu.buildFromTemplate([
+  {
+    label: "Menu",
+    submenu: [
+      {
+        label: "app.label.close",
+        click: () => {
+          ipcMain?.emit("context-menu-command", "menu-item-1");
+        },
+      },
+      {
+        label: "app.label.close",
+        click: () => {
+          ipcMain?.emit("context-menu-command", "menu-item-1");
+        },
+      },
+    ],
+  },
+  {
+    label: "DevTools",
+    click: () => {
+      ipcMain?.emit("context-menu-command", "dev-tools");
+    }
+  },
+]);
+Menu.setApplicationMenu(menu);
 
 app.disableHardwareAcceleration();
 const mainWindow = () => {
@@ -19,20 +45,30 @@ const mainWindow = () => {
       contextIsolation: true,
       preload: path.join(__dirname, "/preload.babel.ts"),
     },
-    minWidth: 600,
-    frame: false,
-    minHeight: 800,
+    minWidth: 800,
+    minHeight: 600,
     // opacity: 0.5
   });
   win.loadURL(AppUrl);
+  ipcMain.on("context-menu-command", (args, type) => {
+    if(typeof args === "string" && args === "dev-tools") {
+      if(win.webContents.isDevToolsOpened()){
+        win.webContents.closeDevTools()
+      } else {
+        win.webContents.openDevTools()
+      }
+    }
+  });
+  ipcMain.on("window-reload-app", () => {
+    win.reload()
+  })
   return win;
 };
 app.whenReady().then(() => {
   // @ts-expect-error
   app.allowRendererProcessReuse = true;
-  if (app.dock) {
-    app.dock.bounce();
-  }
+
+  
   mainWindow();
   app.on("activate", () => {
     if (BrowserWindow.getAllWindows().length === 0) {