+++ preload.babel.ts
@@ -4,7 +4,8 @@ const { contextBridge, ipcRenderer } = require("electron");
 
 // console.log(__ipcMain)/
 contextBridge.exposeInMainWorld('electronApp', {
-    onDownload: () => ipcRenderer.send('download-started')
+    onDownload: () => ipcRenderer.send('download-started'),
+    reloadWindow: () => ipcRenderer.send("window-reload-app")
 })
 // if(window){
 //   window.ipcMain = ipcMain;